decouple kernel check configs + misc refactor improvements
