+++ .github/scripts/versioning.py
@@ -69,7 +69,7 @@ def ask_yes_no_question(question):
     return answer
 
 
-def list_files_in_a_component(component, afr_path, exclude_dirs=[], ext_filter=['.c', '.h']):
+def list_files_in_a_component(component, afr_path, exclude_dirs=[], ext_filter=['.c', '.h'], exclude_hidden=True):
     '''
     Returns a list of all the files in a component.
     '''
@@ -77,19 +77,17 @@ def list_files_in_a_component(component, afr_path, exclude_dirs=[], ext_filter=[
     search_path = os.path.join(afr_path, component)
 
     for root, dirs, files in os.walk(search_path, topdown=True):
-        dirs[:] = [d for d in dirs if d not in exclude_dirs]
-
-        # Do not include hidden files and folders.
-        dirs[:] = [d for d in dirs if not d[0] == '.']
-        files = [f for f in files if not f[0] == '.']
+        # Current root is an excluded dir so skip
+        if root in exclude_dirs:
+            continue
 
         for f in files:
-            if ext_filter != None:
-                ext = '.' + f.split('.')[-1]
-                if ext in ext_filter:
+            if exclude_hidden and f[0] == '.':
+                continue
+
+            if (ext_filter is None
+                or ext_filter is not None and os.path.splitext(f)[-1] in ext_filter):
                     list_of_files.append(os.path.join(os.path.relpath(root, afr_path), f))
-            else:
-                list_of_files.append(os.path.join(os.path.relpath(root, afr_path), f))
 
     return list_of_files
 
@@ -223,9 +221,10 @@ def update_freertos_version_macros(path_macrofile, major, minor, build):
     with open(path_macrofile, 'w', newline='') as macro_file:
         macro_file.write(macro_file_content)
 
-def update_version_number_in_freertos_component(component, root_dir, old_version_prefix_list, new_version, verbose=False):
+def update_version_number_in_freertos_component(component, root_dir, old_version_prefix_list, new_version,
+                                                verbose=False, exclude_hidden=True):
     assert isinstance(old_version_prefix_list, list), 'Expected a list for arg(old_version_prefix_list)'
-    component_files = list_files_in_a_component(component, root_dir, ext_filter=None)
+    component_files = list_files_in_a_component(component, root_dir, ext_filter=None, exclude_hidden=exclude_hidden)
     version_numbers = defaultdict(list)
     n_updated = 0
